Refactor too long dive.py into separate files
